trader rebuild just begin

+++ model/log.go
@@ -11,8 +11,8 @@ import (
 
 // Log struct
 type Log struct {
-	ID           uint64 `gorm:"primary_key;AUTO_INCREMENT"`
-	TraderID     uint   `gorm:"index"`
+	ID           int64 `gorm:"primary_key;AUTO_INCREMENT"`
+	TraderID     int64 `gorm:"index"`
 	Timestamp    int64
 	ExchangeType string `gorm:"type:varchar(50)"`
 	Type         int    // [-1"error", 0"info", 1"profit", 2"buy", 3"sell", 4"cancel", 5"long", 6"short", 7"long_close", 8"short_close"]
@@ -26,7 +26,7 @@ type Log struct {
 
 // Logger struct
 type Logger struct {
-	TraderID     uint
+	TraderID     int64
 	ExchangeType string
 }
